requirements.txt

flask==2.3.3
flask-cors==4.0.0
langchain==0.0.267
chromadb==0.4.13
openai==0.27.8
beautifulsoup4==4.12.2
python-dotenv==1.0.0
langchain-google-genai==2.1.2